AG - Added customer billing tests
